% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/use_gui.R
\name{use_gui}
\alias{use_gui}
\title{use_gui}
\usage{
use_gui(name = system.file("apps/main.Rmd", package = "sbgui"))
}
\arguments{
\item{name}{a name}
}
\description{
A one liner.
}
\details{
Here more details
}
\author{
Reinhard SImon
}

